DEV: Upgrade react-native to 0.71.7 (#186)
Fixes a build issue with latest XCode and this error:
```
FBReactNativeSpec/FBReactNativeSpec.h:2063:46: 'value' is unavailable
```

Fixes a react-native-siri-shortcut issue using a patch. This also broke builds with an error:

```
react-native-siri-shortcut/ios/RNSSAddToSiriButtonViewManager.m:45:70: 'new' is unavailable

INUIAddVoiceShortcutButton *button = [INUIAddVoiceShortcutButton new];
```

https://github.com/Gustash/react-native-siri-shortcut/issues/131

The issue isn't fixed yet upstream, so we're going to be using a patch for now.